Numerous AA50 bug fixes
git-svn-id: https://svn.code.sf.net/p/triplea/code/trunk/triplea@2295 a4c711aa-641f-0410-879d-9b464e52b682

+++ src/games/strategy/triplea/ui/ProductionRepairPanel.java
@@ -271,7 +271,7 @@ public class ProductionRepairPanel extends JPanel
     public class Rule extends JPanel
     {
         private ScrollableTextField m_text = new ScrollableTextField(0, Integer.MAX_VALUE);
-        private int m_cost;         
+        private float m_cost;         
         private RepairRule m_rule;
         private String m_terr;
 
@@ -314,7 +314,7 @@ public class ProductionRepairPanel extends JPanel
             setBorder(new EtchedBorder());
         }
 
-        int getCost()
+        float getCost()
         {
             return m_cost;
         }
